Catch the MalformedURLException that can be thrown when redirecting
If the request URL is malformed, and we are trying to parse it for the redirect, just return a 404 not found error.

This will prevent the exception from showing up in the log. I haven't seen this due to normal use, but something with the network security scanning that is regularly done at my place of work was causing such an exception.